chore: update model for new data

+++ its_jointprobability/models/prodslda.py
@@ -76,10 +76,12 @@ class ProdSLDA(Model):
 
         if batch is not None:
             batch_size = len(batch)
-            docs = docs[list(batch)]
-            labels = labels[list(batch)] if labels is not None else None
+            docs = docs[list(batch)].float()
+            labels = labels[list(batch)].float() if labels is not None else None
         else:
             batch_size = num_full_data
+            docs = docs.float()
+            labels = labels.float()
 
         labels_plate = pyro.plate("labels", self.label_size, dim=-2)
         docs_plate = pyro.plate("documents", docs.shape[0], dim=-1)
@@ -152,12 +154,15 @@ class ProdSLDA(Model):
     ):
         num_full_data = docs.shape[0]
 
+
         if batch is not None:
             batch_size = len(batch)
-            docs = docs[list(batch)]
-            labels = labels[list(batch)] if labels is not None else None
+            docs = docs[list(batch)].float()
+            labels = labels[list(batch)].float() if labels is not None else None
         else:
             batch_size = num_full_data
+            docs = docs.float()
+            labels = labels.float()
 
         labels_plate = pyro.plate("labels", self.label_size, dim=-2)
         docs_plate = pyro.plate("documents", docs.shape[0], dim=-1)
